+++ clown_sort/__init__.py
@@ -15,6 +15,7 @@ from clown_sort.config import Config
 from clown_sort.files.image_file import ImageFile
 from clown_sort.files.pdf_file import PdfFile
 from clown_sort.files.sortable_file import SortableFile
+from clown_sort.sort_selector import process_file
 from clown_sort.util.filesystem_helper import (files_in_dir, is_image, is_pdf,
      set_timestamp_based_on_screenshot_filename)
 
@@ -23,7 +24,14 @@ def sort_screenshots():
     Config.configure()
 
     for image in screenshot_paths(Config.screenshots_dir):
-        image.sort_file()
+        if Config.manual_sort:
+            if not isinstance(image, ImageFile):
+                print(f"'{image.file_path}' is not suitable for manual sort, skipping...")
+                continue
+
+            process_file(image)
+        else:
+            image.sort_file()
 
 
 def set_screenshot_timestamps():